refactoring, moved unfinished unit tests out into temporary directory to bring them in after tests are written

+++ spec/classes/rules/common/sec_aide_spec.rb
@@ -1,30 +1,54 @@
 require 'spec_helper'
 
+enforce_options = [true, false]
+
 describe 'security_baseline::rules::common::sec_aide' do
   on_supported_os.each do |os, os_facts|
-    context "on #{os}" do
-      let(:facts) { os_facts }
-      let(:params) do
-        {
-          'enforce' => true,
-          'message' => 'aide package',
-          'loglevel' => 'warning',
-        }
-      end
 
-      it { is_expected.to compile }
-      it do
-        is_expected.to contain_package('aide')
-          .that_notifies('Exec[aidedb]')
-        is_expected.to contain_exec('aidedb')
-          .with(
-            command: 'aide --init',
+    enforce_options.each do |enforce|
+      context "on #{os}" do
+        let(:facts) { 
+          os_facts.merge(
+            'security_baseline' => {
+              'aide' => {
+                'version' =>  '6.1.2',
+                'status' => 'not installed',
+              }
+            }
           )
-          .that_notifies('Exec[rename_aidedb]')
-        is_expected.to contain_exec('rename_aidedb')
-          .with(
-            command: 'mv /var/lib/aide/aide.db.new.gz /var/lib/aide/aide.db.gz',
-          )
+        }
+        let(:params) do
+          {
+            'enforce' => enforce,
+            'message' => 'aide package',
+            'log_level' => 'warning',
+          }
+        end
+
+        it { is_expected.to compile }
+        it do
+          if enforce
+            is_expected.to contain_package('aide')
+              .that_notifies('Exec[aidedb]')
+            is_expected.to contain_exec('aidedb')
+              .with(
+                command: 'aide --init',
+              )
+              .that_notifies('Exec[rename_aidedb]')
+            is_expected.to contain_exec('rename_aidedb')
+              .with(
+                command: 'mv /var/lib/aide/aide.db.new.gz /var/lib/aide/aide.db.gz',
+              )
+          else
+            is_expected.to contain_echo('aide')
+              .with(
+                'message'  => 'aide package',
+                'loglevel' => 'warning',
+                'withpath' => false,
+              )
+            
+          end
+        end
       end
     end
   end
